+++ instrumentation/opentelemetry-instrumentation-grpc/tests/test_client_interceptor.py
@@ -20,10 +20,6 @@ from tests.protobuf import (  # pylint: disable=no-name-in-module
 import opentelemetry.instrumentation.grpc
 from opentelemetry import trace
 from opentelemetry.instrumentation.grpc import GrpcInstrumentorClient
-from opentelemetry.sdk.metrics.export.aggregate import (
-    MinMaxSumCountAggregator,
-    SumAggregator,
-)
 from opentelemetry.test.test_base import TestBase
 
 from ._client import (
@@ -38,9 +34,7 @@ from ._server import create_test_server
 class TestClientProto(TestBase):
     def setUp(self):
         super().setUp()
-        GrpcInstrumentorClient().instrument(
-            exporter=self.memory_metrics_exporter
-        )
+        GrpcInstrumentorClient().instrument()
         self.server = create_test_server(25565)
         self.server.start()
         self.channel = grpc.insecure_channel("localhost:25565")
@@ -49,56 +43,8 @@ class TestClientProto(TestBase):
     def tearDown(self):
         super().tearDown()
         GrpcInstrumentorClient().uninstrument()
-        self.memory_metrics_exporter.clear()
         self.server.stop(None)
         self.channel.close()
-
-    def _verify_success_records(self, num_bytes_out, num_bytes_in, method):
-        # pylint: disable=protected-access,no-member
-        self.channel._interceptor.controller.tick()
-        records = self.memory_metrics_exporter.get_exported_metrics()
-        self.assertEqual(len(records), 3)
-
-        bytes_out = None
-        bytes_in = None
-        duration = None
-
-        for record in records:
-            if record.instrument.name == "grpcio/client/duration":
-                duration = record
-            elif record.instrument.name == "grpcio/client/bytes_out":
-                bytes_out = record
-            elif record.instrument.name == "grpcio/client/bytes_in":
-                bytes_in = record
-
-        self.assertIsNotNone(bytes_out)
-        self.assertEqual(bytes_out.instrument.name, "grpcio/client/bytes_out")
-        self.assertEqual(bytes_out.labels, (("rpc.method", method),))
-
-        self.assertIsNotNone(bytes_in)
-        self.assertEqual(bytes_in.instrument.name, "grpcio/client/bytes_in")
-        self.assertEqual(bytes_in.labels, (("rpc.method", method),))
-
-        self.assertIsNotNone(duration)
-        self.assertEqual(duration.instrument.name, "grpcio/client/duration")
-        self.assertSequenceEqual(
-            sorted(duration.labels),
-            [
-                ("rpc.grpc.status_code", grpc.StatusCode.OK.name),
-                ("rpc.method", method),
-                ("rpc.system", "grpc"),
-            ],
-        )
-
-        self.assertEqual(type(bytes_out.aggregator), SumAggregator)
-        self.assertEqual(type(bytes_in.aggregator), SumAggregator)
-        self.assertEqual(type(duration.aggregator), MinMaxSumCountAggregator)
-
-        self.assertEqual(bytes_out.aggregator.checkpoint, num_bytes_out)
-        self.assertEqual(bytes_in.aggregator.checkpoint, num_bytes_in)
-
-        self.assertEqual(duration.aggregator.checkpoint.count, 1)
-        self.assertGreaterEqual(duration.aggregator.checkpoint.sum, 0)
 
     def test_unary_unary(self):
         simple_method(self._stub)
@@ -113,8 +59,6 @@ class TestClientProto(TestBase):
         self.check_span_instrumentation_info(
             span, opentelemetry.instrumentation.grpc
         )
-
-        self._verify_success_records(8, 8, "/GRPCTestServer/SimpleMethod")
 
         self.assert_span_has_attributes(
             span,
@@ -140,10 +84,6 @@ class TestClientProto(TestBase):
             span, opentelemetry.instrumentation.grpc
         )
 
-        self._verify_success_records(
-            8, 40, "/GRPCTestServer/ServerStreamingMethod"
-        )
-
         self.assert_span_has_attributes(
             span,
             {
@@ -166,10 +106,6 @@ class TestClientProto(TestBase):
         # Check version and name in span's instrumentation info
         self.check_span_instrumentation_info(
             span, opentelemetry.instrumentation.grpc
-        )
-
-        self._verify_success_records(
-            40, 8, "/GRPCTestServer/ClientStreamingMethod"
         )
 
         self.assert_span_has_attributes(
@@ -198,10 +134,6 @@ class TestClientProto(TestBase):
             span, opentelemetry.instrumentation.grpc
         )
 
-        self._verify_success_records(
-            40, 40, "/GRPCTestServer/BidirectionalStreamingMethod"
-        )
-
         self.assert_span_has_attributes(
             span,
             {
@@ -212,64 +144,9 @@ class TestClientProto(TestBase):
             },
         )
 
-    def _verify_error_records(self, method):
-        # pylint: disable=protected-access,no-member
-        self.channel._interceptor.controller.tick()
-        records = self.memory_metrics_exporter.get_exported_metrics()
-        self.assertEqual(len(records), 3)
-
-        bytes_out = None
-        errors = None
-        duration = None
-
-        for record in records:
-            if record.instrument.name == "grpcio/client/duration":
-                duration = record
-            elif record.instrument.name == "grpcio/client/bytes_out":
-                bytes_out = record
-            elif record.instrument.name == "grpcio/client/errors":
-                errors = record
-
-        self.assertIsNotNone(bytes_out)
-        self.assertIsNotNone(errors)
-        self.assertIsNotNone(duration)
-
-        self.assertEqual(errors.instrument.name, "grpcio/client/errors")
-        self.assertSequenceEqual(
-            sorted(errors.labels),
-            sorted(
-                (
-                    (
-                        "rpc.grpc.status_code",
-                        grpc.StatusCode.INVALID_ARGUMENT.name,
-                    ),
-                    ("rpc.method", method),
-                    ("rpc.system", "grpc"),
-                )
-            ),
-        )
-        self.assertEqual(errors.aggregator.checkpoint, 1)
-
-        self.assertSequenceEqual(
-            sorted(duration.labels),
-            sorted(
-                (
-                    ("error", "true"),
-                    ("rpc.method", method),
-                    ("rpc.system", "grpc"),
-                    (
-                        "rpc.grpc.status_code",
-                        grpc.StatusCode.INVALID_ARGUMENT.name,
-                    ),
-                )
-            ),
-        )
-
     def test_error_simple(self):
         with self.assertRaises(grpc.RpcError):
             simple_method(self._stub, error=True)
-
-        self._verify_error_records("/GRPCTestServer/SimpleMethod")
 
         spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans), 1)
@@ -282,7 +159,6 @@ class TestClientProto(TestBase):
         with self.assertRaises(grpc.RpcError):
             client_streaming_method(self._stub, error=True)
 
-        self._verify_error_records("/GRPCTestServer/ClientStreamingMethod")
         spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans), 1)
         span = spans[0]
@@ -293,8 +169,6 @@ class TestClientProto(TestBase):
     def test_error_unary_stream(self):
         with self.assertRaises(grpc.RpcError):
             server_streaming_method(self._stub, error=True)
-
-        self._verify_error_records("/GRPCTestServer/ServerStreamingMethod")
 
         spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans), 1)
@@ -307,43 +181,9 @@ class TestClientProto(TestBase):
         with self.assertRaises(grpc.RpcError):
             bidirectional_streaming_method(self._stub, error=True)
 
-        self._verify_error_records(
-            "/GRPCTestServer/BidirectionalStreamingMethod"
-        )
-
         spans = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(spans), 1)
         span = spans[0]
         self.assertIs(
             span.status.status_code, trace.status.StatusCode.ERROR,
-        )
-
-
-class TestClientNoMetrics(TestBase):
-    def setUp(self):
-        super().setUp()
-        GrpcInstrumentorClient().instrument()
-        self.server = create_test_server(25565)
-        self.server.start()
-        self.channel = grpc.insecure_channel("localhost:25565")
-        self._stub = test_server_pb2_grpc.GRPCTestServerStub(self.channel)
-
-    def tearDown(self):
-        super().tearDown()
-        GrpcInstrumentorClient().uninstrument()
-        self.memory_metrics_exporter.clear()
-        self.server.stop(None)
-
-    def test_unary_unary(self):
-        simple_method(self._stub)
-        spans = self.memory_exporter.get_finished_spans()
-        self.assertEqual(len(spans), 1)
-        span = spans[0]
-
-        self.assertEqual(span.name, "/GRPCTestServer/SimpleMethod")
-        self.assertIs(span.kind, trace.SpanKind.CLIENT)
-
-        # Check version and name in span's instrumentation info
-        self.check_span_instrumentation_info(
-            span, opentelemetry.instrumentation.grpc
         )