+++ instrumentation/opentelemetry-instrumentation-requests/tests/test_requests_integration.py
@@ -22,7 +22,6 @@ import opentelemetry.instrumentation.requests
 from opentelemetry import context, propagators, trace
 from opentelemetry.instrumentation.requests import RequestsInstrumentor
 from opentelemetry.sdk import resources
-from opentelemetry.sdk.util import get_dict_as_key
 from opentelemetry.test.mock_textmap import MockTextMapPropagator
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace.status import StatusCode
@@ -91,26 +90,6 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.check_span_instrumentation_info(
             span, opentelemetry.instrumentation.requests
         )
-
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.flavor": "1.1",
-                "http.method": "GET",
-                "http.status_code": "200",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
-                self.assertGreaterEqual(view_data.aggregator.current.sum, 0)
 
     def test_name_callback(self):
         def name_callback(method, url):
@@ -289,23 +268,6 @@ class RequestsIntegrationTestBase(abc.ABC):
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
 
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
-
     mocked_response = requests.Response()
     mocked_response.status_code = 500
     mocked_response.reason = "Internal Server Error"
@@ -323,23 +285,6 @@ class RequestsIntegrationTestBase(abc.ABC):
             span.attributes, {"http.method": "GET", "http.url": self.URL}
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
-
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
 
     mocked_response = requests.Response()
     mocked_response.status_code = 500
@@ -364,23 +309,6 @@ class RequestsIntegrationTestBase(abc.ABC):
             },
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
-        self.assertIsNotNone(RequestsInstrumentor().meter)
-        self.assertEqual(len(RequestsInstrumentor().meter.instruments), 1)
-        recorder = list(RequestsInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.status_code": "500",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
 
     @mock.patch("requests.adapters.HTTPAdapter.send", side_effect=Exception)
     def test_requests_basic_exception(self, *_, **__):