+++ instrumentation/opentelemetry-instrumentation-urllib/tests/test_urllib_integration.py
@@ -29,7 +29,6 @@ from opentelemetry.instrumentation.urllib import (  # pylint: disable=no-name-in
     URLLibInstrumentor,
 )
 from opentelemetry.sdk import resources
-from opentelemetry.sdk.util import get_dict_as_key
 from opentelemetry.test.mock_textmap import MockTextMapPropagator
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace.status import StatusCode
@@ -116,26 +115,6 @@ class RequestsIntegrationTestBase(abc.ABC):
         self.check_span_instrumentation_info(
             span, opentelemetry.instrumentation.urllib
         )
-
-        self.assertIsNotNone(URLLibInstrumentor().meter)
-        self.assertEqual(len(URLLibInstrumentor().meter.instruments), 1)
-        recorder = list(URLLibInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.flavor": "1.1",
-                "http.method": "GET",
-                "http.status_code": "200",
-                "http.url": "http://httpbin.org/status/200",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
-                self.assertGreaterEqual(view_data.aggregator.current.sum, 0)
 
     def test_name_callback(self):
         def name_callback(method, url):
@@ -325,24 +304,6 @@ class RequestsIntegrationTestBase(abc.ABC):
             },
         )
         self.assertEqual(span.status.status_code, StatusCode.ERROR)
-        self.assertIsNotNone(URLLibInstrumentor().meter)
-        self.assertEqual(len(URLLibInstrumentor().meter.instruments), 1)
-        recorder = list(URLLibInstrumentor().meter.instruments.values())[0]
-        match_key = get_dict_as_key(
-            {
-                "http.method": "GET",
-                "http.status_code": "500",
-                "http.url": "http://httpbin.org/status/500",
-                "http.flavor": "1.1",
-            }
-        )
-        for key in recorder.bound_instruments.keys():
-            self.assertEqual(key, match_key)
-            # pylint: disable=protected-access
-            bound = recorder.bound_instruments.get(key)
-            for view_data in bound.view_datas:
-                self.assertEqual(view_data.labels, key)
-                self.assertEqual(view_data.aggregator.current.count, 1)
 
     def test_requests_basic_exception(self, *_, **__):
         with self.assertRaises(Exception):