+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#301](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/301))
 - `opentelemetry-instrumentation-asgi` Return header values using case insensitive keys
   ([#308](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/308))
+- Remove metrics from all instrumentations
+  ([#312](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/312))
 
 ## [0.17b0](https://github.com/open-telemetry/opentelemetry-python-contrib/releases/tag/v0.17b0) - 2021-01-20